third_party/blink/web_tests/platform/linux-chrome/external/wpt/web-animations/timing-model/timelines/timelines-expected.txt

This is a testharness.js-based test.
[FAIL] Timeline time increases once per animation frame in an iframe
  assert_greater_than: Timeline time within an iframe increases between animation frames expected a number greater than 0 but got 0
Harness: the test ran to completion.
